Compose: Add `useDelayedLoading` hook
