+++ packages/compose/src/index.js
@@ -43,6 +43,7 @@ export { default as useAsyncList } from './hooks/use-async-list';
 export { default as useWarnOnChange } from './hooks/use-warn-on-change';
 export { default as useDebounce } from './hooks/use-debounce';
 export { default as useDebouncedInput } from './hooks/use-debounced-input';
+export { default as useDelayedLoading } from './hooks/use-delayed-loading';
 export { default as useThrottle } from './hooks/use-throttle';
 export { default as useMergeRefs } from './hooks/use-merge-refs';
 export { default as useRefEffect } from './hooks/use-ref-effect';
